+++ src/textwatermark/__init__.py
@@ -1,5 +1,5 @@
 '''TextWatermark'''
-__version__ = '0.1.2'
+__version__ = '0.1.3'
 
 from .main import TextWatermark
 from .template import WMTemplate